requirements.txt
boto3==1.4.1
python-twitter==3.3
